refactor: Update methods of BeneficiaryRepository and switch class type to sealed

+++ src/Infrastructure/Persistence/Repositories/BeneficiaryRepository.cs
@@ -4,9 +4,8 @@ using Microsoft.EntityFrameworkCore;
 
 namespace Infrastructure.Persistence.Repositories;
 
-public class BeneficiaryRepository(AppDbContext context) : IBeneficiaryRepository
+public sealed class BeneficiaryRepository(AppDbContext context) : IBeneficiaryRepository
 {
-
     public async Task<Beneficiary> CreateAsync(Beneficiary beneficiary)
     {
         await context.Beneficiaries.AddAsync(beneficiary);
@@ -17,24 +16,13 @@ public class BeneficiaryRepository(AppDbContext context) : IBeneficiaryRepositor
     public async Task DeleteAsync(Guid id)
     {
         var beneficiary = await context.Beneficiaries.FirstOrDefaultAsync(b => b.Id == id);
-
-        if (beneficiary == null)
-        {
-            throw new ArgumentNullException("The Beneficiary with id = " + id + " does not exist");
-        }
-        context.Beneficiaries.Remove(beneficiary);
+        if (beneficiary != null) context.Beneficiaries.Remove(beneficiary);
         await context.SaveChangesAsync();
     }
 
     public async Task<Beneficiary> UpdateAsync(Guid id, Beneficiary beneficiaryToUpdate)
     {
         var beneficiary = await context.Beneficiaries.FirstOrDefaultAsync(b => b.Id == id);
-        
-        if (beneficiary == null)
-        {
-            throw new ArgumentNullException("The Beneficiary with id = " + id + " does not exist");
-        }
-        
         context.Beneficiaries.Update(beneficiaryToUpdate);
         await context.SaveChangesAsync();
         return beneficiaryToUpdate;
